+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 	- `sort`
 * Added support for common table expressions to the query builder.
 * Added `Query::forCompiler()` method that can be used to add dialect specific SQL to queries.
+* Added `Query::selectRaw()` method.
 * Added `Query::whereColumn()` method.
 * Added query builder date helpers:
 	- `Query::whereDate()`